Update halo2 dependency to v2023_04_20
Resolve #1366

+++ zkevm-circuits/src/evm_circuit/execution/balance.rs
@@ -125,7 +125,7 @@ impl<F: Field> ExecutionGadget<F> for BalanceGadget<F> {
 
         let (_, is_warm) = block.rws[step.rw_indices[4]].tx_access_list_value_pair();
         self.is_warm
-            .assign(region, offset, Value::known(F::from(is_warm)))?;
+            .assign(region, offset, Value::known(F::from(is_warm as u64)))?;
 
         let code_hash = block.rws[step.rw_indices[5]].account_value_pair().0;
         self.code_hash
